pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agelclaw"
version = "3.1.0"
description = "Self-evolving AI agent with persistent memory, skills, and multi-provider routing"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Stefanos Drakos", email = "sdrakos@agel.ai" }]

dependencies = [
    "claude-agent-sdk>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0", "openai-agents>=0.1.0"]
outlook = ["msal>=1.28.0"]
embeddings = ["sqlite-vec>=0.1.0", "openai>=1.0.0"]
all = ["agelclaw[openai,outlook,embeddings]"]

[project.scripts]
agelclaw = "agelclaw.cli_entry:main"
agelclaw-mem = "agelclaw.mem_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agelclaw"]

[tool.hatch.build.targets.wheel.force-include]
"src/agelclaw/data" = "agelclaw/data"
